Sublime-Syntax Changes
Added support for pre-processor directives to the sublime-syntax file,
and a corresponding .json file storing the replacements e.g.
'{{@KEYWORD}}' -> 'keyword.tex'. Using this approach, the syntax file
has been updated and tidied up. (In particular, it now worries a bit
less about structure - e.g. it will now highlight '\item' as if it were
valid within an itemize, whether or not it occurs inside one. Before, it
would only do so inside an itemize. Seeing as how TeX syntax can be
changed pretty much arbitrarily on the fly, it makes sense to worry less
about this sort of thing.)