Mini Overhaul
Mostly changes to the way in which we can reference variables in the
syntax.json within the actual syntax files (and at the same time, we
change the json file to have a more logical structure). In this way we
have a more robust and modular approach. Accordingly, the syntax files
have changed a bit, and the python file that translates them into legal
sublime-syntax files has become a bit smarter. Finally (I think?) we
have added some MetaPost support (i.e. the syntax gets highlighted
nicely :D), and refactored some of the ConTeXt.sublime-syntax file.