Adding some snippets
We add some simple snippets in addition to the start/stop one already
here. Perhaps these snippets could be improved upon, I do like that they
remind you of the relevant syntax though. Also some housekeeping, and we
add a cool Lua thing - now <Toggle Comment> (default 'Ctrl - /') inserts
a '%' in plain text, but knows to instead insert a '--' when writing Lua
inside ConTeXt (or it does '--[[ ... ]]' if you use <Toggle Block
Comment>). Furthermore, we give up, for the moment, on keeping track of
what functions are defined in the Lua namespace in ConTeXt. (Should we
go through and add them all in? There are a lot!)