added command completions
We have created an `EventListener` which does two things.

(1) it adds all ~4000 commands in the ConTeXt interface to the
completions list. These are generated automatically from the XML
interface files provided in the ConTeXt (version 2016.05.17 19:20)
distribution in TeX live.

(2) Whenever one of those commands is typed out, a popup appears showing
the full documentation for that command (as described by the XML files).