Better command information
We've done a better job at parsing the interface XML files. Well, that
is true, but furthermore we fix a couple of errors/omissions in the XML
too.

As such, we now handle certain argument types correctly e.g. <cd:resolve
name="argument-content-list"/> which comes up in e.g. \doifelsecommon.
We also handle <cd:instances> ... </cd:instances> properly, which gives
us a couple of \...makeup commands we had previously ignored. Finally
when the file attribute is missing, we report it as unknown for now.
(That said, this seems to mostly apply to various XML commands so we
could possibly track down the source file without too much difficulty;
we don't bother for now.)