Misc small updates
A change to the structure of commands.json, in light of the fact that
\bar, \but, \nop, \txt, \underbar and \overbar have different
definitions in mulitple files... and a corresponding change to
completions.py.

A handful of additions/improvements to the syntax tests.

Some modifications to the main syntax file, mainly concerning how Lua is
handled. Lua in TeX is tricky, this change fixes an edge case with
strings at least.