Better user settings.
The main changes are how settings are defined, and updates in line with
this.

Currently you can set up multiple profiles in the user settings (as we
illustrate with an example in the default settings) and choose between
them by setting the `current_profile` key. In this way we aim to
smoothly handle multiple ConTeXt installations on the same machine. We
also modify `Main.sublime-menu` to have options for the default and user
settings.

To go along with this change, we define a `ContexttoolsProfileSelector`
command and hook it up with the `ConTeXtTools.sublime-commands` file.
This command acts as a short-cut to opening the settings file up and
manually changing the `current_profile` key. We also tweak the
`buildPDF` and `completions` scripts to understand the new settings
structure, and start putting the big JSON files into their own
`interface` folder.