References, misc
Refactored the existing python plugins. We now have some common
functionality in the `scripts` folder, and the main plugins import these
files.

Building on this, we have some new functionality for working with
references. The keys `<profile>/references/command_regex` and
`<profile>/references/reference_regex` can be used in the user settings
to control what is detected as a reference (things like
`[eq:pythagoras]`) and at what times to prompt the user to select a
reference (e.g. after typing `\eqref`).

Alterations to the log and build results syntaxes.

Tweaked the snippets and added `\startTABLE ... \stopTABLE` and `<?lua
... ?>` ones.

Hooked up some of the commands to the ST API, i.e. put in entries into
the `sublime-commands` file.