More settings, updated README, separate MetaFun
Update to the README in line with recent developments.

Separated MetaFun off into its own syntax file, which uses the MetaPost
syntax as a base. Tweaks to the MetaPost syntax file itself. Added
syntax tests for the new MetaFun file, as well as one for the `ConTeXt
Log` syntax.

Renaming `lmx` to the more descriptive `template` in various places.

Changes to settings, they now support a bigger variety of options (as
described in the README).

More functionality lumped into the `common` script, including a
`reload_settings` function which understands the updated way of handling
settings. Corresponding changes to `references`, `completions` and
`buildPDF` scripts, as well as other tweaks here and there to all three.