Main syntax rewrite, tweak other bits accordingly
Rewrite of the main syntax file with correct scope names as a priority.
Having done this, it becomes easy to use the symbol list for headings
(chapter/part/section etc.), and we set this up in the symbol folder.

Changes to the python scripts that rely on the syntax file. BuildPDF can
now change the command name to use, defaults to "context".

Other misc changes.

Updated the README.