Markup tweaks
Handling both styles---`{... \emph ...}` and `...
\emph{...}`---correctly is tricky. We want them to work nested, e.g.
`\emph{... \bold{...} ... }`, and also handle them as switches (e.g.
`{... \emph ... \bold ...}`), and they should work anywhere, except that
they *definitely* shouldn't work in cases like
`\setuphead[chapter][style=\emph]`.

So there's lots of fiddly edge cases. This implementation works alright,
but comes with the trade-off that it doesn't understand having e.g.
`\emph` and its argument `{...}` on separate lines. Another thing that
this version doesn't do is understand that in, say, `{... \em ... \em
...}` the code after the second `\em` should be back in plain style.
Similarly with `{\emph{... \emph{...} ...}`, and also in the case `{...
\em ... \tf ...}` it doesn't understand the switch `\tf`. We could add
this stuff in, might not be worth the effort though.