Syntax goodies
Put in some more metadata in the syntax, e.g.
`meta.other.control.word.context` now means exactly a control sequence,
and `meta.other.title.context` means exactly a heading name.

Continuing to experiment with treating control sequences like `\leq`
differently in math mode.

Put in experimental table/markup handling. Trying to support markup like
`{\bold ...}` (font switch isolated to a group) and `\bold{...}` (font
applied to entire subgroup), and now adding in the case `\NC \bold ...
\NC` (or `\bTD \bold ... \eTD`) in a table (font switch isolated to a
group, this time the group is implicit) continues to be a messy
business.

This table + markup juggling act is maybe not worth the effort, as
behind-the-scenes the code is getting ugly, but the end result *is*
pretty slick.