Refactoring
Adding to/tweaking the metadata that the main syntax file attaches to
ConTeXt files. Updated the syntax test to reflect the current state of
the syntax.

On the scripting side of things, the new metadata allows for simplifying
many bits and pieces. At the same time the code gets cleaned up a tad.

Changed the `.sublime-menu` to use the `edit_settings` command, which
seems like a better way of handling things.