More optimizing
The `last_command_in_view` function is called every time the buffer gets
modified, so it needs to be very quick. On larger files (with 1000s of
lines) the previous implementation became painfully slow. The new
version seems to be a lot faster.

Also changed the sorting on pop-ups a tiny bit. Should be slightly
quicker, and also smarter.