Optimisations
`last_command_in_view` (and `last_command_with_args_in_view`, which has
been made redundant and thus removed) seemed to be acting up on large
(say, ~5000 lines) files. That is, becoming incredibly slow.
Speculation: it was looking *all* the way back to the start of the file,
even though that should (?) be impossible. Now there is a default limit
to how far it will look back. (Hopefully 200 turns out to be a
reasonable default.) This seems to have resolved the issue.