Async Builder
Finally, we figured out how to thread the builder! We run `context` via
subprocess, in its own private thread using `threading`. Also we
discovered the `CREATE_NO_WINDOW` flag for windows.

As to the finer details of things, we use locks and state variables to
try and make the builder thread-safe.