Syntax corrections, new settings system
Some small fixes to some `sublime-syntax` files. Besides that, starting
on a new way of managing the settings, instead of the profile system.
The motivation for this change is that the profile system, which is
pretty good overall, is bad at making small changes; you have to make a
whole separate profile if you want to, say, turn on the `--arrange` flag
for compiling, or change the colours for the pop-ups.

So the new approach is (will be) more straightforward: you can go into
the settings in a nice GUI, using `quick_panels` provided by the ST API,
and change whatever you want at will. It should also be able to emulate
"profiles" as before: there will be a "scheme" feature (might rename
better): schemes are simply a groups of settings, e.g. a scheme that
dictates to use the interface file & engine for some-such version of
context. Then you'll be able to select that scheme in the GUI, which
will apply all the settings it contains. In line with all these changes,
of course, the structure of the settings files will be changing some.

That's the plan, anyway. This commit it just an initial attempt at
implementing it. `commands.py` is where the GUI stuff is located, and
the other python files have been made aware of the new system. The GUI
is partially functional at this moment.