Finished settings GUI
Nice! The system looks pretty robust, and is easy to use. We could in
future extend it, to be more powerful. But for the time being, I'm very
happy with it: you can do a lot just in the menus, although occasionally
you will need to edit the `sublime-settings` directly, e.g. to define
new schemes.