Mini-overhaul
Mostly, how pop-ups are done behind-the-scenes has been improved a
lot. Now there is an emphasis on HTML, since (mini-)HTML is the
target format. What was \`parsing.py\` has been broken up into the
two scripts \`interface_reading.py\` and \`interface_writing.py\`,
for the two halves of the job. All the code for this has been
redone from scratch, so it should be cleaner, more robust, and
more extensible.

We've also renamed \`common.py\` to \`utilities.py\` and some of
the code in there has been redone/improved upon. In particular,
there is some HTML-related code in there, for tasks that the
standard library couldn't do out of the box. This HTML code is a
bit crude, but it should do the job well enough. Another
important addition is \`locate\`, which calls \`mtxrun --locate
...\` via \`subprocess\` in order to locate whatever ConTeXt
related files we need. (Note: I *think* that this is the proper
way of doing things in ConTeXt, compared with \`kpathsea\` in a
LaTeX system.)

Besides this, there are one or two tweaks to the
\`.sublime-syntax\` files. Also the \`.sublime-settings\` file
has changed a little bit, mostly to try obtain a more logical
structure, and also to accomodate the different scripts. The
\`.sublime-commands\` file has changed, taking out some commands
that no longer exis (for the moment). Of particular interest is
taking out the command for generating an interface, we now do
it automatically if needs be.

Some other bits and bobs got tweaked or renamed.