pop-ups on hover, lots of refactoring
Added more metadata into the ConTeXt sublime syntax, to be eventually used by an auto formatter. Also added some comments, and minor additions (e.g. supporting `\directlua 0 {...}`).

Added another build variation for MetaPost: this one calls `context` directly on the MetaPost file. As this script is clever enough to distinguish between a TeX and a MetaPost, it will run MetaFun on the file.

Split `utilities.py` up into logically grouped modules. As a result, various calls to `utilities.some_function` have been modified.

Completions have new goodies. The original auto-complete style has been refactored but remains basically the same. In addition there is an on-hover version of the same functionality. You can choose to have any combination of both these features in the new setting options. At the same time, took out the `key=value` feature for now. I plan to add it back in later, in a more complete form, but as the backend has changed we leave it for now.

Fixed a bug with `file_links` sometimes creating an error message after successfully opening a file

Took out the regex-based reference handling. The plan is to gradually add metadata for references into the syntax file (as we have already started to do), and do some analysis at the python end. This should leave us with a more automatic and powerful reference system.

Minor changes here and there as a consequence of the aforementioned ones.