snippets, minihtml, scopes
Renamed/reorganized the existing snippets. Added some new ones, including project ones which autocomplete with the current file name. Need to update the README to reflect these changes.

Tweaked the symbol files in light of the changed scope names.

The ConTeXt builder now defaults to a simple call `context $file` (using the builtin `exec` target in Sublime Text). Might as well put that in there. For our custom builder, that is now covered in some variants on the builder (\<kbd\>Ctrl\</kbd\>+\<Shift\>+\<kbd\>B\</kbd\> to bring up the list of builders by default).

In the main ConTeXt syntax file, we have done a big reorganisation. The functionality is mostly the same, with `\protect` taken out for now (to be put added back in better) and some other bits and pieces added in. The scope names have changed, to be more in line the Sublime Text recommendations. And there are some comments here and there.

Added some Lua builders, to ease the use of LuaTeX as a standalone Lua interpreter.

Regarding pop-ups, things are a work in progress, some bugs to work out.

Forgot to add some CSS tags; documented a new one.

`file_links.py` has some weird bugs going on. More investiation needed.