snippets, builders, misc
Added some new snippets, and renamed some existing ones.

The LuaTeX and MetaPost builders have been stripped down to the bare essentials. The built-in builder can handle all of the book-keeping needed, so just provide a simple wrapper around it. Also fits in more with the ST way of doing things.

The main ConTeXt builder, however, cannot be simplified in the same way. Tidy it up some for now.