Update README, add autogenerate setting, better copy popup
Updated the README with various things. The `ConTeXt.sublime-syntax` file is now exactly the output of the YAML Macros builder. Added and documented a `pop_ups/try_generate_on_demand` option. Changed the functionality of the `pop_ups/show_copy_pop_up` option: the HTML option was useful for debugging, but for anything else we'd rather just copy the plain text. (We leave in the option of turning on the HTML if needs be.) Changed `file_as_slug` to hash the file path instead of what it currently does. We *should* really handle the possibility of a hash collision, but for now we just ignore it.