Embed syntaxes, clean up
Switch to using the `embed` action in our `.sublime-syntax` files, where possible. We do have to do some tricks with `embed` to get the behaviour we want, but it works, and this enables us to add in some freebies e.g. `\startXML ... \stopXML` no problem.

On the other hand, we've taken out `\unprotect ... \protect` completely, undecided what to do about that. I was hoping there might be some nice YAML Macros trick to get what we want, but I don't think that's possible.