Syntax fixes, buffers, key=val completions
Some fixes for the syntax.

Added in a basic completion thing for buffers: we give them a special scope in the syntax file, and using that we can do a kind of autocomplete for them (like with references). It's rudimentary, just a little experiment really for now.

Added back in the `key=value` style completions for commands. Intended to be a very simple approach, a more sophisticated version might e.g. mind which argument we are currently in and act accordingly. Comes with an option to turn it off.