Settings class, citations, clean up
Put our common methods into a proper class, and then subclass it everywhere needed. This is *much* better to my eyes. (Refactor things accordingly). Also, go ahead and pull out whatever class-level variables we can.

Starting work on some citation stuff. We have the `.lua` format taken care of so far: we can rely on the existence of a Lua interpreter in the form of LuaTeX, so simply load the Lua file and pass it along to Python in a format it understands (I chose JSON). The `.xml` format should be pretty straightforward too, just use Python's built-in XML parsing abilities. The tricky one will be the traditional `.bib` format. I'm thinking of using Lua's LPEGs to handle it, as we have Lua(TeX) as an option for free.

General Python tidying up. Took out some stuff we aren't using anymore.

Added an option `script/timeout` for the script runner.