Citations
Continuing work on the citation system. In addition to the `.lua` parser we now have a `.bib` and `.xml` parser: the full trifecta. The BibTeX parser is build out of LPEGs, which have been a pleasure to work with. The XML parser simply uses the Python standard library. All three parsers produce a JSON-ish Python `dict`.