Tweak interface stuff, clean up codebase some.
Tweak `is_visible`, so as to not accidentally hide everything outside of a ConTeXt file. Style `cd:threearguments` as `\...#1#2#3`, and similarly for `cd:twoarguments` and `cd:oneargument`. Some general refactoring, hopefully not breaking too much. Prefix it whenever we write to the ST terminal, to make it clear it's us saying that.